requirements.txt
kubernetes
pyyaml
jinja2